Revert "Fix app installation: enable default linkapps for other apps while keeping 1Password in system Applications"
This reverts commit 16d27e7b5a4c74b38a75a147ce8c11972ca439f7.

+++ libraries/home-manager/systems/darwin/default.nix
@@ -24,20 +24,13 @@ in {
       app="''${applications[$i]}"
       rm -rf "$BASE_DIR/$app"
 
-      if [ -d "${apps}/Applications/$app" ]; then
-        if [[ "$app" = "1Password.app" ]]; then
-          # 1Password는 시스템 Applications에 설치
-          rm -rf "/Applications/$app"
-          $DRY_RUN_CMD cp ''$VERBOSE_ARG -fHRL "${apps}/Applications/$app" "/Applications"
-          $DRY_RUN_CMD chmod ''$VERBOSE_ARG -R +w "/Applications/$app"
-        else
-          # 다른 앱들은 사용자 Applications에 설치
-          $DRY_RUN_CMD cp ''$VERBOSE_ARG -fHRL "${apps}/Applications/$app" "$BASE_DIR"
-          $DRY_RUN_CMD chmod ''$VERBOSE_ARG -R +w "$BASE_DIR/$app"
-        fi
+      if [ -d "${apps}/Applications/$app" ] && [[ "$app" = "1Password.app" ]]; then
+        rm -rf "/Applications/$app"
+        $DRY_RUN_CMD cp ''$VERBOSE_ARG -fHRL "${apps}/Applications/$app" "/Applications"
+        $DRY_RUN_CMD chmod ''$VERBOSE_ARG -R +w "/Applications/$app"
       fi
     done
   '';
 
-  # disabledModules = ["targets/darwin/linkapps.nix"];  # 기본 linkapps 모듈 사용
+  disabledModules = ["targets/darwin/linkapps.nix"];
 }
